+++ keanu-project/src/main/java/io/improbable/keanu/tensor/TensorShape.java
@@ -8,6 +8,7 @@ import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.List;
 
+
 public class TensorShape {
 
     private long[] shape;