+++ keanu-project/src/main/java/io/improbable/keanu/tensor/dbl/DoubleTensor.java
@@ -100,6 +100,10 @@ public abstract class DoubleTensor implements NumberTensor<Double, DoubleTensor>
         return new ScalarDoubleTensor(scalarValue);
     }
 
+    static DoubleTensor vector(double... values) {
+        return DoubleTensor.create(values, values.length);
+    }
+
     /**
      * @param dimension the dimension along which toStack are stacked
      * @param toStack   an array of DoubleTensor's of the same shape